feat: Recently viewed files are now saved in iDB
This commit adds a new feature to the app. The recently viewed files are now saved in IndexedDB. This means that the user can close the app and reopen it and the files will still be there. This is a huge improvement over the previous version of the app.
Dexie is used to interact with the iDB.
Also Syntax highlighting is now done using the react-syntax-highlighter package.